move, add: Providerの移動, ReduxProviderの追加
